+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED
+
+- Added "follow_corrections" API flag (#23)
+
 ## 2.4.0 (2022-08-01)
 
 - Set Heroku runtime to python-3.10.5 (#21)