+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 
 - Added "follow_corrections" API flag (#23)
+- Added configuration to host the project on fly.io
+- Updated the address of the sample installation to https://linguee-api.fly.dev
 
 ## 2.4.0 (2022-08-01)